scraper: number recognition for protected fields
maoyan.com protects certain numeric fields from being scrapped in the following way:

- picked some unused unicode characters
- generate a font file that encodes 0-9 to the picked characters

In this method, the displayed information (say, rating) is human-readable, but if you copy it to somewhere else, it will appear as ., and it is not recognized by computer. 

With the added code, scraper.py can now recognize the number values of the protected fields by comparing the strokes of the font to a known font.